Added option to shoot down hanggliders flying in restricted airspace
Added a /area_flak command to declare areas as closed airspace. Any
hanggliders attempting to fly in such an area will get shot down
within two seconds, which will lead to a massive loss of health to the
occupant (death if flying higher than 6 meters) and the loss of the
hangglider. The closed airspace behaviour depends on the areas mod.